refcator split out resolvers, typedefs, add card games prework
