+++ src/services/cardgames/cardGameResolvers.ts
@@ -5,31 +5,23 @@ const cardGameResolvers = {
   Query: {
     getAllCardGames: async () => {
       return await CardGameService.getAllCardGames();
-    },
-    getCardGameById: async (_: any, args: { id: string }) => {
-      return await CardGameService.getCardGameById(parseInt(args.id));
+  },
+    getCardGameById: async (_: any, { id }: { id: string }) => {
+      return await CardGameService.getCardGameById(parseInt(id));
     },
   },
   Mutation: {
     createCardGame: async (
       _: any,
-      { name, price, description, img, categoryId, stock, preorder, rrp }: any
+      { name, description, img, categoryId }: { name: string; description: string; img: any; categoryId: number }
     ) => {
-      return await CardGameService.createCardGame(
-        name,
-        price,
-        description,
-        img,
-        categoryId,
-        stock,
-        preorder,
-        rrp
-      );
+      return await CardGameService.createCardGame(name, description, img, categoryId);
     },
-    updateCardGame: async (_: any, { id, name, price, description, img, categoryId, stock, preorder, rrp }: any) => {
-      return await CardGameService.updateCardGame(
-        id, name, price, description, img, categoryId, stock, preorder, rrp
-      );
+    updateCardGame: async (
+      _: any,
+      { id, name, description, img, categoryId }: { id: string; name: string; description: string; img: any; categoryId: number }
+    ) => {
+      return await CardGameService.updateCardGame(id, name, description, img, categoryId);
     },
     deleteCardGame: async (_: any, { id }: { id: string }) => {
       return await CardGameService.deleteCardGame(id);
@@ -39,11 +31,6 @@ const cardGameResolvers = {
     category: async (parent: any) => {
       return await prisma.category.findUnique({
         where: { id: parent.categoryId },
-      });
-    },
-    stock: async (parent: any) => {
-      return await prisma.stock.findUnique({
-        where: { productId: parent.id },
       });
     },
     img: async (parent: any) => {
